% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_parser.R
\name{as.vector.question}
\alias{as.vector.question}
\title{S3 method for converting a question to a vector}
\usage{
\method{as.vector}{question}(question)
}
\description{
S3 method for converting a question to a vector
}
