% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy.load.R
\name{easy.load}
\alias{easy.load}
\title{Easy Load}
\usage{
easy.load(
  survey.path = NA,
  experiment.data = NA,
  experiment.file.names.vector = NA,
  experiment.files.structure
)
}
\arguments{
\item{survey.path}{}

\item{experiment.data}{}

\item{experiment.file.names.vector}{}

\item{experiment.files.structure}{}
}
\value{
data.frame
}
\description{
Easy Load
}
