% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.experiment.data.R
\name{load.experiment.data}
\alias{load.experiment.data}
\title{Load Experiment Data}
\usage{
load.experiment.data(
  experiment.file.names,
  label.structure,
  merge.dataframe = NA
)
}
\arguments{
\item{experiment.file.names}{A vector of experiment filenames}

\item{merge.dataframe}{The result of this function can be directly added to a existing dataframe (e.g. survey data).
The default value is NA.}

\item{label.structur}{A vector of labels that represents the structur of the saved experiment value}
}
\value{
data.frame
}
\description{
This functions loads experiment data from a Psytoolkit experiment
Therefore, a vector of filenames and a vector of labels which represents the
structure of the experiment file have to be provided
}
