% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.experiment.data.R
\name{load.experiment}
\alias{load.experiment}
\title{Load Experiment}
\usage{
load.experiment(experiment.file.names, label.structure, merge.dataframe = NA)
}
\arguments{
\item{experiment.file.names}{A vector of experiment file names}

\item{merge.dataframe}{The result of this function can be directly added to a existing data frame (e.g. survey data).
The default value is NA.}

\item{label.structur}{A vector of labels that represents the structure of the saved experiment value}
}
\value{
data.frame
}
\description{
This function loads experiment data from a PsyToolkit experiment
Therefore, a vector of file names and a vector of labels which represents the
structure of the experiment file have to be provided.
The experiment data has to be located in the `psy_data/experiment_data` directory.
Otherwise, loading the experiment will fail.
}
