% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.survey.R
\name{load.survey}
\alias{load.survey}
\title{Load Survey}
\usage{
load.survey(filename)
}
\arguments{
\item{filename}{File name of the .ZIP file that has been downloaded from PsyToolkit}
}
\value{
data.frame
}
\description{
This functions loads the answers of a PsyToolkit survey from the specified .ZIP file.
This file will be unzipped in the folder `psy_data`. Do not remove this folder for further use such as reading experiment data.
}
