% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.survey.R
\name{parse.survey}
\alias{parse.survey}
\title{Parse survey file}
\usage{
parse.survey()
}
\description{
This function returns a survey opject that holds information on the survey.
The survey data is retrieved from survey.txt
