% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{question}
\alias{question}
\title{Question object (S3)
This is a S3 object that represents a survey question.}
\usage{
question()
}
\description{
Question object (S3)
This is a S3 object that represents a survey question.
}
