% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.status.R
\name{recode.status}
\alias{recode.status}
\title{Recode Status
This function recodes the status variable to custom values.}
\usage{
recode.status(status_vector, correct = 1, error = 0, timeout = NA)
}
\arguments{
\item{status_vector}{}

\item{correct}{}

\item{error}{}

\item{timeout}{}
}
\value{
numeric vector
}
\description{
Recode Status
This function recodes the status variable to custom values.
}
