% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_status.R
\name{recode_status}
\alias{recode_status}
\title{Recode Status}
\usage{
recode_status(status_vector, correct = 1, error = 0, timeout = NA)
}
\arguments{
\item{status_vector}{}

\item{correct}{}

\item{error}{}

\item{timeout}{}
}
\value{
numeric vector
}
\description{
Recode Status
}
