% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{survey}
\alias{summary.survey}
\title{Survey object (S3)}
\usage{
survey()

\method{summary}{survey}(object)
}
\arguments{
\item{object}{survey object}
}
\description{
This function gives a summary of a survey.
It displays information on each question, such as label, question, and options.
}
\section{Functions}{
\itemize{
\item \code{survey()}: survey object

}}
