% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.experiment.R
\name{load.experiment}
\alias{load.experiment}
\title{Load Experiment}
\usage{
load.experiment(experiment.file.names, label.structure, merge.dataframe = NA)
}
\arguments{
\item{experiment.file.names}{vector of experiment file names}

\item{label.structure}{vector of labels that represent the structure of a experiment data file}

\item{merge.dataframe}{The result of this function can be directly added to a existing data frame (e.g. survey data). The default value is NA.}
}
\value{
data.frame
}
\description{
This function loads experiment data from a PsyToolkit experiment
The experiment files have to be located in the `psy_data/experiment_data` directory.
}
