% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.keys.R
\name{recode.keys}
\alias{recode.keys}
\title{Recode Keys}
\usage{
recode.keys(vector, keys)
}
\arguments{
\item{vector}{A vector of key values}

\item{keys}{A vector of keys. Those keys have to be in the same order than
in your experiment.}
}
\value{
numeric
}
\description{
Recode Keys
}
