% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.status.R
\name{recode.status}
\alias{recode.status}
\title{Recode Status}
\usage{
recode.status(status_vector, correct = 1, error = 0, timeout = NA)
}
\arguments{
\item{status_vector}{Numeric vector of statuses. This vector will be recoded.}

\item{correct}{Numeric value which indicates a correct response.}

\item{error}{Numeric value which indicates an error.}

\item{timeout}{Numeric value which indicates a timeout.}
}
\value{
numeric
}
\description{
This function recodes the status variable to custom values.
}
