% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.status.R
\name{recode.status}
\alias{recode.status}
\title{Recode Status}
\usage{
recode.status(status_vector, correct = 1, error = 0, timeout = NA)
}
\arguments{
\item{status_vector}{numeric vector of statuses. This vector will be recoded.}

\item{correct}{numeric value which indicates a correct response.}

\item{error}{numeric value which indicates an error.}

\item{timeout}{numeric value which indicates a timeout.}
}
\value{
numeric
}
\description{
In PsyToolkit experiments, a status of 1 represents a correct and 2 an incorrect answer. 3 represents a time out.
With recode.status(), status codes can be recoded to custom values.
}
\examples{
recode.status(vector, correct = 1, error = 0, timeout = -99)
}
